--- student-projects/Assaf_Project/Module-III/class-035-14-12-2023/my_work/class-project.md
@@ -1,7 +1,0 @@
-1. Add option to toggle todos
-   - Todo can be done/pending
-   - Done todos are displayed with strikethrough
-2. Add "clear done todos" button
-3. Add buttons to change view: all todos/pending todos/done todos
-
-* No javascript in the client
