--- student-projects/Assaf_Project/Module-III/class-035-14-12-2023/my_work/todos.json
@@ -1,22 +1,0 @@
-[
- {
-  "id": "66359c18-f289-4ee2-9093-f119190339a2",
-  "text": "what is love",
-  "isDone": true
- },
- {
-  "id": "96b31e2b-f1de-45ea-ac74-792615688576",
-  "text": "baby don't hurt me",
-  "isDone": false
- },
- {
-  "id": "27a527fa-0519-4a83-8b9f-db9295b1f73d",
-  "text": "don't hurt me",
-  "isDone": false
- },
- {
-  "id": "21890867-3140-4885-b0d6-15728df2e0f0",
-  "text": "no more",
-  "isDone": false
- }
-]
