--- student-projects/Assaf_Project/Module-III/class-035-14-12-2023/my_work/tsconfig.json
@@ -1,13 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "ESNext",
-        "lib": ["ESNext"],
-        "moduleResolution": "NodeNext",
-        "module": "NodeNext",
-        "strictNullChecks": true,
-        "noImplicitAny": true,
-        "useUnknownInCatchVariables": true,
-        "esModuleInterop": true
-    },
-    "include": ["./*.ts"]
-}
