--- student-projects/Assaf_Project/Module-III/class-035-14-12-2023/my_work/views/todos.ejs
@@ -1,60 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Todos</title>
-    <link rel="stylesheet" href="/style.css" />
-</head>
-
-<body>
-    <header class="--padding">
-        <h1>Todos</h1>
-        <nav>
-            <a class="btn" href="/">All</a>
-            <a class="btn" href="/pending">Pending</a>
-            <a class="btn" href="/done">Done</a>
-        </nav>
-    </header>
-    <main class="--padding">
-        <form action="/addTodo" method="post">
-            <input type="text" name="todo" autofocus />
-            <button class="btn" type="submit">Add</button>
-        </form>
-        <ul class="todo-list">
-            <% for (let todo of todos) { %>
-                <li<%- todo.isDone ? ` class="flexed --done" ` : ` class="flexed" ` %>>
-                    <form action="/toggle" method="post">
-                        <input type="hidden" name="id" value="<%= todo.id %>">
-                        <button class="clearBtn" type="submit">
-                            <%= todo.isDone ? "❌" : "✅" %>
-                        </button>
-                        <%=todo.text%>
-                    </form>
-                    <form action="/deleteTodo" method="post">
-                        <input type="hidden" name="delete" value="<%= todo.id %>">
-                        <button class="clearBtn" type="submit">🗑️</button>
-                    </form>
-                    </li>
-                    <% } %>
-        </ul>
-        <p>Todos count: <%= todos.length %>
-        </p>
-        <menu class="flexed --gap">
-            <form action="/resetTodos" method="post">
-                <button class="btn" type="submit">Reset todos</button>
-            </form>
-            <form action="/clearDone" method="post">
-                <button class="btn" type="submit">Clear Done</button>
-            </form>
-        </menu>
-    </main>
-    <article class="--padding">
-        <p>
-            <%= visits %> "people" has viewed this page
-        </p>
-    </article>
-</body>
-
-</html>
