+++ student-projects/Assaf_Project/Module-III/class-035-14-12-2023/views/todos.ejs
@@ -9,28 +9,52 @@
 </head>
 
 <body>
-    <h1>Todos</h1>
-    <form action="/addTodo" method="post">
-        <input type="text" name="todo" autofocus />
-        <button type="submit">Add</button>
-    </form>
-    <ul>
-        <% for (let todo of todos) { %>
-            <li class="<%= todo.completed ? 'completed' : 'pending' %>">
-                <%= todo %>
-                <form action="/toggle" method="post">
-                    <button type="submit">toggle</button></form>
-            </li>
-            <% } %>
-    </ul>
-    <p>Todos count: <%= todos.length %>
-    </p>
-    <form action="/resetTodos" method="post">
-        <button type="submit">Reset todos</button>
-    </form>
-    <p>
-        <%= visits %> "people" has viewed this page
-    </p>
+    <header class="--padding">
+        <h1>Todos</h1>
+        <nav>
+            <a class="btn" href="/">All</a>
+            <a class="btn" href="/pending">Pending</a>
+            <a class="btn" href="/done">Done</a>
+        </nav>
+    </header>
+    <main class="--padding">
+        <form action="/addTodo" method="post">
+            <input type="text" name="todo" autofocus />
+            <button class="btn" type="submit">Add</button>
+        </form>
+        <ul class="todo-list">
+            <% for (let todo of todos) { %>
+                <li<%- todo.isDone ? ` class="flexed --done" ` : ` class="flexed" ` %>>
+                    <form action="/toggle" method="post">
+                        <input type="hidden" name="id" value="<%= todo.id %>">
+                        <button class="clearBtn" type="submit">
+                            <%= todo.isDone ? "❌" : "✅" %>
+                        </button>
+                        <%=todo.text%>
+                    </form>
+                    <form action="/deleteTodo" method="post">
+                        <input type="hidden" name="delete" value="<%= todo.id %>">
+                        <button class="clearBtn" type="submit">🗑️</button>
+                    </form>
+                    </li>
+                    <% } %>
+        </ul>
+        <p>Todos count: <%= todos.length %>
+        </p>
+        <menu class="flexed --gap">
+            <form action="/resetTodos" method="post">
+                <button class="btn" type="submit">Reset todos</button>
+            </form>
+            <form action="/clearDone" method="post">
+                <button class="btn" type="submit">Clear Done</button>
+            </form>
+        </menu>
+    </main>
+    <article class="--padding">
+        <p>
+            <%= visits %> "people" has viewed this page
+        </p>
+    </article>
 </body>
 
 </html>
